+++ ddGetDocuments_snippet.php
@@ -1,13 +1,13 @@
 <?php
 /**
  * ddGetDocuments
- * @version 1.1 (2020-07-05)
+ * @version 1.2 (2020-10-09)
  * 
  * @see README.md
  * 
  * @link https://code.divandesign.biz/modx/ddgetdocuments
  * 
- * @copyright 2015–2020 DD Group {@link http://www.DivanDesign.biz }
+ * @copyright 2015–2020 DD Group {@link https://DivanDesign.biz }
  */
 
 //The snippet must return an empty string even if result is absent
@@ -59,11 +59,6 @@ $offset =
 	isset($offset) ?
 	$offset :
 	0
-;
-$orderBy =
-	isset($orderBy) ?
-	$orderBy :
-	'`id` ASC'
 ;
 $filter =
 	isset($filter) ?
@@ -122,6 +117,12 @@ $extendersParams =
 if(class_exists($dataProviderClass)){
 	//Prepare provider params
 	$providerParams = \ddTools::encodedStringToArray($providerParams);
+	
+	//Backward compatibility with <= 1.1
+	if (isset($orderBy)){
+		$providerParams['orderBy'] = $orderBy;
+	}
+	
 	//Prepare extender params
 	$extendersParams = (object) \ddTools::encodedStringToArray($extendersParams);
 	//Prepare outputter params
@@ -150,11 +151,13 @@ if(class_exists($dataProviderClass)){
 	}
 	
 	//Make sure orderBy and filter looks like SQL
-	$orderBy = str_replace(
-		$fieldDelimiter,
-		'`',
-		$orderBy
-	);
+	if (!empty($providerParams['orderBy'])){
+		$providerParams['orderBy'] = str_replace(
+			$fieldDelimiter,
+			'`',
+			$providerParams['orderBy']
+		);
+	}
 	$filter = str_replace(
 		$fieldDelimiter,
 		'`',
@@ -165,7 +168,6 @@ if(class_exists($dataProviderClass)){
 		'snippetParams' => [
 			'offset' => $offset,
 			'total' => $total,
-			'orderBy' => $orderBy,
 			'filter' => $filter
 		],
 		'providerParams' => $providerParams,
@@ -195,6 +197,9 @@ if(class_exists($dataProviderClass)){
 		
 		//Overwrite the snippet parameters with the result of applying them to the current extender
 		$input->snippetParams = $extenderObject->applyToSnippetParams($input->snippetParams);
+		
+		//Overwrite the data provider parameters with the result of applying them to the current extender
+		$input->providerParams = $extenderObject->applyToDataProviderParams($input->providerParams);
 	}
 	
 	$dataProviderObject = new $dataProviderClass($input);