+++ src/Extender/Pagination/Extender.php
@@ -2,8 +2,6 @@
 namespace ddGetDocuments\Extender\Pagination;
 
 
-use ddGetDocuments\DataProvider\DataProviderOutput;
-
 class Extender extends \ddGetDocuments\Extender\Extender {
 	private
 		/**
@@ -45,7 +43,7 @@ class Extender extends \ddGetDocuments\Extender\Extender {
 	
 	/**
 	 * __construct
-	 * @version 1.2.1 (2019-03-19)
+	 * @version 1.2.2 (2020-10-02)
 	 * 
 	 * @param $params {stdClass|arrayAssociative}
 	 */
@@ -53,32 +51,19 @@ class Extender extends \ddGetDocuments\Extender\Extender {
 		//Call base constructor
 		parent::__construct($params);
 		
-		if($this->pageTpl != ''){
-			$this->pageTpl = \ddTools::$modx->getTpl((string) $this->pageTpl);
-		}
-		
-		if($this->currentPageTpl != ''){
-			$this->currentPageTpl = \ddTools::$modx->getTpl((string) $this->currentPageTpl);
-		}
-		
-		if($this->wrapperTpl != ''){
-			$this->wrapperTpl = \ddTools::$modx->getTpl((string) $this->wrapperTpl);
-		}
-		
-		if($this->nextTpl != ''){
-			$this->nextTpl = \ddTools::$modx->getTpl((string) $this->nextTpl);
-		}
-		
-		if($this->nextOffTpl != ''){
-			$this->nextOffTpl = \ddTools::$modx->getTpl((string) $this->nextOffTpl);
-		}
-		
-		if($this->previousTpl != ''){
-			$this->previousTpl = \ddTools::$modx->getTpl((string) $this->previousTpl);
-		}
-		
-		if($this->previousOffTpl != ''){
-			$this->previousOffTpl = \ddTools::$modx->getTpl((string) $this->previousOffTpl);
+		foreach (
+			[
+				'pageTpl',
+				'currentPageTpl',
+				'wrapperTpl',
+				'nextTpl',
+				'nextOffTpl',
+				'previousTpl',
+				'previousOffTpl'
+			] as
+			$tplName
+		){
+			$this->{$tplName} = \ddTools::$modx->getTpl((string) $this->{$tplName});
 		}
 		
 		$this->pageIndex =
@@ -90,7 +75,7 @@ class Extender extends \ddGetDocuments\Extender\Extender {
 	
 	/**
 	 * applyToSnippetParams
-	 * @version 2.0 (2020-03-11)
+	 * @version 2.0.1 (2020-10-02)
 	 * 
 	 * @param $snippetParams {stdClass}
 	 * 
@@ -100,7 +85,10 @@ class Extender extends \ddGetDocuments\Extender\Extender {
 		//If “total” is set then we need to override “offset” according to the current page index
 		if(isset($snippetParams->total)){
 			$snippetParams->offset =
-				($this->pageIndex - 1) *
+				(
+					$this->pageIndex -
+					1
+				) *
 				$snippetParams->total
 			;
 		}
@@ -112,13 +100,13 @@ class Extender extends \ddGetDocuments\Extender\Extender {
 	
 	/**
 	 * applyToOutput
-	 * @version 1.2 (2020-06-22)
+	 * @version 1.2.2 (2020-10-02)
 	 * 
 	 * @param $dataProviderOutput {\ddGetDocuments\DataProvider\DataProviderOutput}
 	 * 
 	 * @return {string}
 	 */
-	public function applyToOutput(DataProviderOutput $dataProviderOutput){
+	public function applyToOutput(\ddGetDocuments\DataProvider\DataProviderOutput $dataProviderOutput){
 		$result = '';
 		
 		//Check to prevent division by zero
@@ -180,7 +168,8 @@ class Extender extends \ddGetDocuments\Extender\Extender {
 						'text' => $pageChunk,
 						'data' => [
 							'url' =>
-								$urlPrefix.$this->pageIndexRequestParamName .
+								$urlPrefix .
+								$this->pageIndexRequestParamName .
 								'=' .
 								$pageIndex
 							,
@@ -214,7 +203,10 @@ class Extender extends \ddGetDocuments\Extender\Extender {
 										$urlPrefix .
 										$this->pageIndexRequestParamName .
 										'=' .
-										($this->pageIndex - 1)
+										(
+											$this->pageIndex -
+											1
+										)
 									)
 								,
 								'totalPages' => $pagesTotal
@@ -231,7 +223,10 @@ class Extender extends \ddGetDocuments\Extender\Extender {
 										$urlPrefix .
 										$this->pageIndexRequestParamName .
 										'=' .
-										($this->pageIndex + 1)
+										(
+											$this->pageIndex +
+											1
+										)
 									)
 								,
 								'totalPages' => $pagesTotal