+++ src/Extender/Search/Extender.php
@@ -2,8 +2,6 @@
 namespace ddGetDocuments\Extender\Search;
 
 
-use ddGetDocuments\DataProvider\DataProviderOutput;
-
 class Extender extends \ddGetDocuments\Extender\Extender {
 	private
 		$currentQuery = ''